requirements.txt
# Web Framework
fastapi
uvicorn[standard]
motor
python-dotenv
pydantic[email]
dnspython

# Testing Framework
robotframework
robotframework-requests
requests

# Optional: สำหรับเทส async ในอนาคต
httpx